Better setup and changed from atLib to ctools - Cleaned up premake scripts for all the projects and broke some parts down into components which can be called from different lua files. - ctools is supposed to be more of a collection of independent classes to help with certain tasks. At lib was becoming more of a framework and quite cumbersome. The design of it wasn't exactly planned. - Added 3rdParty requirements needed for Fractal which were in atLib (TODO: add repos as submodules and add a step to the Fractal project to build/link each 3rdParty repo). - Added generate-project.bat in the root directory which displays the options for different project formats. So far only added on windows. I haven't even attempted to build Fractal on Linux yet, but I plan to. When that happens I'll add better project gen support for Linux. - Updated mouse inputs to be relative to the window that they belong to. The global mouse input is in screen coords. Need to add a global/relative mouse pos option for the Mouse input devices (I'll add a task for that on github) - Fixed some warnings that show up in vs2019 - Added Window::GetFocusedWindow() to return a pointer to the Fractal window with keyboard focus, mouse focus or both.

+++ Engine/source/graphics/OpenGL/flGLVertexBuffer.cpp
@@ -133,7 +133,7 @@ namespace flEngine
 
       void* GetNativeResource()
       {
-        return (void*)m_bufferID;
+        return (void*)(uint64_t)m_bufferID;
       }
 
     protected:
